+++ src/main/kotlin/com/cout970/magneticraft/gui/common/blocks/ContainerFirebox.kt
@@ -31,7 +31,7 @@ class ContainerFirebox(player: EntityPlayer, world: World, blockPos: BlockPos) :
         val data = IBD()
         data.setFloat(DATA_ID_BURNING_TIME, tile.burningTime)
         data.setFloat(DATA_ID_MAX_BURNING_TIME, tile.maxBurningTime)
-        data.setLong(DATA_ID_MACHINE_HEAT, tile.heat.heat)
+        data.setDouble(DATA_ID_MACHINE_HEAT, tile.heat.heat)
         return data
     }
 
@@ -39,6 +39,6 @@ class ContainerFirebox(player: EntityPlayer, world: World, blockPos: BlockPos) :
         tile!!
         ibd.getFloat(DATA_ID_BURNING_TIME, { tile.burningTime = it })
         ibd.getFloat(DATA_ID_MAX_BURNING_TIME, { tile.maxBurningTime = it })
-        ibd.getLong(DATA_ID_MACHINE_HEAT, { tile.heat.heat = it })
+        ibd.getDouble(DATA_ID_MACHINE_HEAT, { tile.heat.heat = it })
     }
 }
