+++ src/main/kotlin/com/cout970/magneticraft/tileentity/heat/TileHeatSink.kt
@@ -13,7 +13,7 @@ class TileHeatSink : TileHeatBase() {
 
     val heat = HeatContainer(dissipation = 0.05,
             specificHeat = COPPER_HEAT_CAPACITY * 3,
-            maxHeat = (COPPER_HEAT_CAPACITY * 3 * COPPER_MELTING_POINT).toLong(),
+            maxHeat = COPPER_HEAT_CAPACITY * 3 * COPPER_MELTING_POINT,
             conductivity = DEFAULT_CONDUCTIVITY,
             worldGetter = { this.world },
             posGetter = { this.getPos() })
